Code refactoring and login:
Refactored a lot of code so it's more consistent
Moved some files around so it's more organized
Changed names of config samples, so it's clearer what you should do with them

Added login functionality
Added PHPass
Added messages to various functions (such as delete). For now they're just a message, but will be prettier in the future
Added SQL dump of the structure of the database. If you import this, you should no longer see any SQLState errors