Added some fancy new stuff to the application:
Added a way to go to previous/next report in report.php
Added a filter for the competencies. Currently a select box, might be changed to checkboxes later
Added a way to add a competency group to a role (apparently that wasn't implemented yet, oops)

Round creation changed a lot:
Now asks for min/max for own/other department, based on percentages and asks for confirmation as well
Also asks for confirmation when ending a round
More information added to round_overview.tpl

Forgot to include the README.md into the other commit, actually added it now.