Added multitenancy and confirmations for deleting stuff.
Controllers:
Added delete_<type>_confirmation function
Changed the links to delete functions in index.php
Changed the QueryWriter from the default QueryWriter of RedBean to a custom QueryWriter (see next change)
Added multitenancy_writer.php, which adds the tenant id to every query.
Changed the validation in feedback.php to reflect the changes of feedback_form.tpl
Added register and register_post functions to security.php, which creates a new organisation (tenant)
Added login_google function to security.php, which enabled logging in with a Google account. This isn't completely finished though.

Views:
Added delete_confirmation template, which changes based on the type given
Changed the link in functions.tpl so it goes to the correct url
Added register.tpl, which contains a form for creating a new organisation